fix(www): move Sentry.init to module-level in main.tsx
Sentry.init was being invoked inside a Layer.effectDiscard inside the
Effect runtime layer. Because of Rolldown's chunk-splitting, the entire
runtime + @sentry/react SDK landed in the preloaded simple-markdown-editor
chunk, not the main entry. While the SMDE chunk is modulepreload'd, this
made Sentry init order dependent on that chunk loading first and made the
init invisible to anyone auditing the main bundle.

Moving Sentry.init to module top-level in main.tsx makes it explicit,
guarantees it runs before any user code in the entry, and is the only
plausible fix for any reports that 'Sentry isn't initialized' in the
deployed www.

The Analytics Effect layer is now a pure interface implementation that
calls Sentry.addBreadcrumb/setUser/etc., on the already-initialized SDK.
The gate (SentryAnalyticsLayer vs NoopAnalyticsLayer) is unchanged.

Verified with bun precommit (oxfmt, oxlint, tsgo across all workspaces).
