+++ apps/www/src/services/analytics/index.ts
@@ -1,3 +1,3 @@
 export { NoopAnalyticsLayer } from './noop'
-export { captureException, makeSentryAnalyticsLayer, type SentryAnalyticsOptions } from './sentry'
+export { captureException, SentryAnalyticsLayer } from './sentry'
 export { Analytics, type AnalyticsProperties, identify, page, reset, track } from './service'
